# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Список",
                "operationId": "31bd209f731f96fbf72021f67a262472",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Some title"
                                                    },
                                                    "author_id": {
                                                        "type": "integer",
                                                        "example": 12
                                                    },
                                                    "published_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "1994-09-03"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-11-24T14:38:49.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-11-24T14:38:49.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Создание",
                "operationId": "eaabd90067e93554d9eaa480d67aaae6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/BookStoreRequestSchema"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "title"
                                                },
                                                "author_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "published_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-12-12"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{books}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Просмотр",
                "operationId": "636fe794d50ae966f75b026770731206",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "ID книги",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "author_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "published_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-12-12"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Удалить",
                "operationId": "06cb9f50f2f4b53391738b251a3fea51",
                "parameters": [
                    {
                        "name": "destroy",
                        "in": "path",
                        "description": "ID книги",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Book deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Books"
                ],
                "summary": "Обновление",
                "operationId": "c98516f41dd75bcd88a2c431e2049f27",
                "parameters": [
                    {
                        "name": "patсh",
                        "in": "path",
                        "description": "ID книги",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/BookUpdateRequestSchema"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "title"
                                                },
                                                "author_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "published_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-12-12"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/books/author/{books}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Автор книги",
                "operationId": "32adca6743927f476e416d8bee0e6070",
                "parameters": [
                    {
                        "name": "Book",
                        "in": "path",
                        "description": "ID книги",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Some some name"
                                        },
                                        "bio": {
                                            "type": "string",
                                            "example": "Some some bio"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-24T14:38:49.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2024-11-24T14:38:49.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/books/rentals/{book}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Аренды книги",
                "operationId": "26c41800d892c7d7a06399d62936b75f",
                "parameters": [
                    {
                        "name": "Book",
                        "in": "path",
                        "description": "ID книги",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "book_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "rented_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-11-21"
                                            },
                                            "due_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-12-06"
                                            },
                                            "return_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": null
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-11-24T14:38:49.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-11-24T14:38:49.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BookStoreRequestSchema": {
                "required": [
                    "name",
                    "bio",
                    "published_at"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "title"
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "published_at": {
                        "type": "string",
                        "example": "2024-11-24"
                    }
                },
                "type": "object"
            },
            "BookUpdateRequestSchema": {
                "required": [
                    "name",
                    "bio",
                    "published_at"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "title"
                    },
                    "author_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "published_at": {
                        "type": "string",
                        "example": "2024-11-24"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}